refactor(toolruntime): split runtime.go into focused files
