+++ toolruntime/runtime.go
@@ -3,18 +3,15 @@ package toolruntime
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"log"
 	"net/http"
 	"os"
-	"regexp"
 	"strconv"
 	"strings"
 	"sync/atomic"
 	"time"
-	"unicode/utf8"
 
 	"github.com/google/uuid"
 	"github.com/modelcontextprotocol/go-sdk/mcp"
@@ -24,23 +21,6 @@ import (
 	"github.com/tinfoilsh/confidential-model-router/tokencount"
 	"github.com/tinfoilsh/confidential-model-router/toolcontext"
 	"github.com/tinfoilsh/confidential-model-router/toolprofile"
-)
-
-// Tinfoil-event marker constants. The router emits opt-in progress
-// markers for router-owned tool calls (web search, fetch) by wrapping a
-// JSON payload in `<tinfoil-event>...</tinfoil-event>` tags carried
-// inside the assistant text stream. Every carrier is a spec-conformant
-// frame (chat delta content, response.output_text.delta, or final
-// assistant message text) so OpenAI SDKs that do not recognize the tags
-// simply render them as text. Clients that do recognize the tags strip
-// them with a single regex before rendering. The marker is only emitted
-// when the caller opts in via the tinfoilEventsHeader request header.
-const (
-	tinfoilEventsHeader     = "X-Tinfoil-Events"
-	tinfoilEventsWebSearch  = "web_search"
-	tinfoilEventOpenTag     = "<tinfoil-event>"
-	tinfoilEventCloseTag    = "</tinfoil-event>"
-	tinfoilEventPayloadType = "tinfoil.web_search_call"
 )
 
 const (
@@ -162,489 +142,6 @@ func debugMessagesSummary(messages []any, contentMax int) string {
 		}
 	}
 	return strings.Join(parts, " | ")
-}
-
-// webSearchOptions is the parsed view of the OpenAI web_search_options /
-// Responses `web_search` tool-config surface that we forward to the MCP tools.
-//
-// piiCheck and injectionCheck capture the caller's opt-in for the websearch
-// server's PII and prompt-injection filters. They are tri-state: a nil pointer
-// means the caller did not send the control so the server should fall back to
-// its own default; a non-nil pointer means the caller explicitly requested the
-// filter on or off and the router forwards that decision as a header on the
-// MCP session.
-type webSearchOptions struct {
-	userLocationCountry string
-	allowedDomains      []string
-	excludedDomains     []string
-	searchContextSize   string
-	contentMode         string
-	maxContentChars     int
-	category            string
-	startPublishedDate  string
-	endPublishedDate    string
-	maxAgeHours         *int
-	piiCheck            *bool
-	injectionCheck      *bool
-	// includeActionSources mirrors OpenAI's documented
-	// `include: ["web_search_call.action.sources"]` request flag. When true,
-	// the synthesized web_search_call output items carry an `action.sources`
-	// array in the spec shape `[{type: "url", url: "..."}]`. When false
-	// (the default) the field is omitted so spec-conformant clients are not
-	// billed for data they did not ask for.
-	includeActionSources bool
-}
-
-// normalizedContextSize returns search_context_size folded to its canonical
-// lowercase/trimmed form so tier lookups stay consistent across helpers.
-func (o webSearchOptions) normalizedContextSize() string {
-	return strings.ToLower(strings.TrimSpace(o.searchContextSize))
-}
-
-// maxResults returns the per-search result cap derived from search_context_size.
-// Zero means "let the MCP tool use its own default".
-func (o webSearchOptions) maxResults() int {
-	switch o.normalizedContextSize() {
-	case "low":
-		return searchContextResultsLow
-	case "medium":
-		return searchContextResultsMedium
-	case "high":
-		return searchContextResultsHigh
-	default:
-		return 0
-	}
-}
-
-// tierContentMode returns the content_mode implied by the search_context_size
-// tier. Low and medium favor query-relevant highlights to keep token use
-// predictable; high asks for the full rendered page text. An empty string
-// means "no tier preference, let the MCP server pick its default".
-func (o webSearchOptions) tierContentMode() string {
-	switch o.normalizedContextSize() {
-	case "low", "medium":
-		return contentModeHighlights
-	case "high":
-		return contentModeText
-	default:
-		return ""
-	}
-}
-
-// tierMaxContentChars returns the per-result character budget implied by the
-// tier. Zero means "no tier preference".
-func (o webSearchOptions) tierMaxContentChars() int {
-	switch o.normalizedContextSize() {
-	case "low":
-		return searchContextCharsLow
-	case "medium":
-		return searchContextCharsMedium
-	case "high":
-		return searchContextCharsHigh
-	default:
-		return 0
-	}
-}
-
-// effectiveContentMode picks the caller's explicit content_mode when set and
-// falls back to the tier-implied value otherwise.
-func (o webSearchOptions) effectiveContentMode() string {
-	if mode := strings.ToLower(strings.TrimSpace(o.contentMode)); mode != "" {
-		return mode
-	}
-	return o.tierContentMode()
-}
-
-// effectiveMaxContentChars prefers an explicit positive caller override and
-// falls back to the tier-implied budget.
-func (o webSearchOptions) effectiveMaxContentChars() int {
-	if o.maxContentChars > 0 {
-		return o.maxContentChars
-	}
-	return o.tierMaxContentChars()
-}
-
-// applyToSearchArgs merges forwarded options into a `search` tool-call
-// argument map produced by the model, leaving any model-supplied overrides in
-// place so it can still narrow the request on its own.
-func (o webSearchOptions) applyToSearchArgs(arguments map[string]any) {
-	if arguments == nil {
-		return
-	}
-	if _, set := arguments["max_results"]; !set {
-		if n := o.maxResults(); n > 0 {
-			arguments["max_results"] = n
-		}
-	}
-	if _, set := arguments["content_mode"]; !set {
-		if mode := o.effectiveContentMode(); mode != "" {
-			arguments["content_mode"] = mode
-		}
-	}
-	if _, set := arguments["max_content_chars"]; !set {
-		if n := o.effectiveMaxContentChars(); n > 0 {
-			arguments["max_content_chars"] = n
-		}
-	}
-	if o.userLocationCountry != "" {
-		if _, set := arguments["user_location_country"]; !set {
-			arguments["user_location_country"] = o.userLocationCountry
-		}
-	}
-	if len(o.allowedDomains) > 0 {
-		if _, set := arguments["allowed_domains"]; !set {
-			arguments["allowed_domains"] = toAnySlice(o.allowedDomains)
-		}
-	}
-	if len(o.excludedDomains) > 0 {
-		if _, set := arguments["excluded_domains"]; !set {
-			arguments["excluded_domains"] = toAnySlice(o.excludedDomains)
-		}
-	}
-	if o.category != "" {
-		if _, set := arguments["category"]; !set {
-			arguments["category"] = o.category
-		}
-	}
-	if o.startPublishedDate != "" {
-		if _, set := arguments["start_published_date"]; !set {
-			arguments["start_published_date"] = o.startPublishedDate
-		}
-	}
-	if o.endPublishedDate != "" {
-		if _, set := arguments["end_published_date"]; !set {
-			arguments["end_published_date"] = o.endPublishedDate
-		}
-	}
-	if o.maxAgeHours != nil {
-		if _, set := arguments["max_age_hours"]; !set {
-			arguments["max_age_hours"] = *o.maxAgeHours
-		}
-	}
-}
-
-// applyToFetchArgs merges forwarded options into a `fetch` tool-call argument
-// map. Retrieval-depth, content-mode, category, and date-range knobs are
-// search-only so they are intentionally skipped here; only host-scoped filters
-// are meaningful for a URL-targeted fetch.
-func (o webSearchOptions) applyToFetchArgs(arguments map[string]any) {
-	if arguments == nil {
-		return
-	}
-	if len(o.allowedDomains) > 0 {
-		if _, set := arguments["allowed_domains"]; !set {
-			arguments["allowed_domains"] = toAnySlice(o.allowedDomains)
-		}
-	}
-	if len(o.excludedDomains) > 0 {
-		if _, set := arguments["excluded_domains"]; !set {
-			arguments["excluded_domains"] = toAnySlice(o.excludedDomains)
-		}
-	}
-}
-
-// parseChatWebSearchOptions extracts OpenAI's documented web_search_options
-// fields plus the sibling `filters` block from a Chat Completions request body.
-// Returns the zero value when the block is missing so the caller can forward
-// "no options" without special casing.
-func parseChatWebSearchOptions(body map[string]any) webSearchOptions {
-	opts := webSearchOptions{}
-	if raw, ok := body["web_search_options"].(map[string]any); ok {
-		opts.searchContextSize = stringValue(raw["search_context_size"])
-		opts.userLocationCountry = extractUserLocationCountry(raw["user_location"])
-		opts.allowedDomains = extractAllowedDomains(raw["filters"])
-		opts.excludedDomains = extractExcludedDomains(raw["filters"])
-		applyAdvancedSearchFields(&opts, raw)
-	}
-	if filters, ok := body["filters"].(map[string]any); ok {
-		if len(opts.allowedDomains) == 0 {
-			opts.allowedDomains = extractDomainListFromFilterMap(filters, "allowed_domains")
-		}
-		if len(opts.excludedDomains) == 0 {
-			opts.excludedDomains = extractDomainListFromFilterMap(filters, "excluded_domains")
-		}
-	}
-	opts.piiCheck = parseSafetyOptIn(body["pii_check_options"])
-	opts.injectionCheck = parseSafetyOptIn(body["prompt_injection_check_options"])
-	opts.includeActionSources = parseIncludeActionSources(body)
-	return opts
-}
-
-// parseResponsesWebSearchOptions mirrors parseChatWebSearchOptions for the
-// Responses API, where options live on the `tools[{type:"web_search", ...}]`
-// entries rather than a sibling block.
-func parseResponsesWebSearchOptions(body map[string]any) webSearchOptions {
-	rawTools, _ := body["tools"].([]any)
-	for _, rawTool := range rawTools {
-		tool, _ := rawTool.(map[string]any)
-		if tool == nil {
-			continue
-		}
-		if stringValue(tool["type"]) != "web_search" {
-			continue
-		}
-		opts := webSearchOptions{
-			searchContextSize:   stringValue(tool["search_context_size"]),
-			userLocationCountry: extractUserLocationCountry(tool["user_location"]),
-		}
-		opts.allowedDomains = extractAllowedDomains(tool["filters"])
-		opts.excludedDomains = extractExcludedDomains(tool["filters"])
-		applyAdvancedSearchFields(&opts, tool)
-		opts.piiCheck = parseSafetyOptIn(body["pii_check_options"])
-		opts.injectionCheck = parseSafetyOptIn(body["prompt_injection_check_options"])
-		opts.includeActionSources = parseIncludeActionSources(body)
-		return opts
-	}
-	return webSearchOptions{}
-}
-
-// parseIncludeActionSources reports whether the request opted into OpenAI's
-// documented `web_search_call.action.sources` include flag. The flag lives
-// at the top of the request body as `include: ["..."]` and unlocks the
-// synthesized `action.sources` field on web_search_call output items. It
-// defaults to false so we do not ship extra data to clients that did not
-// ask for it.
-func parseIncludeActionSources(body map[string]any) bool {
-	raw, ok := body["include"].([]any)
-	if !ok {
-		return false
-	}
-	for _, item := range raw {
-		if stringValue(item) == includeActionSourcesFlag {
-			return true
-		}
-	}
-	return false
-}
-
-// includeActionSourcesFlag is OpenAI's documented `ResponseIncludable`
-// enum value that unlocks `web_search_call.action.sources`.
-const includeActionSourcesFlag = "web_search_call.action.sources"
-
-// stripRouterOwnedIncludes removes include entries that only the router
-// knows how to service from the request body it forwards upstream.
-// Upstream inference servers validate `include` against their own schema
-// and reject unknown values, so the router must filter them out after
-// reading them itself. An empty list is removed entirely to avoid
-// shipping `"include":[]`, which some servers still reject.
-func stripRouterOwnedIncludes(body map[string]any) {
-	raw, ok := body["include"].([]any)
-	if !ok {
-		return
-	}
-	filtered := raw[:0]
-	for _, item := range raw {
-		if stringValue(item) == includeActionSourcesFlag {
-			continue
-		}
-		filtered = append(filtered, item)
-	}
-	if len(filtered) == 0 {
-		delete(body, "include")
-		return
-	}
-	body["include"] = filtered
-}
-
-// parseSafetyOptIn reads a caller-provided safety control block such as
-// `pii_check_options` or `prompt_injection_check_options` from a request body.
-//
-// The presence of the key is itself the signal that the caller wants the
-// corresponding filter enabled for this request; the block's contents are
-// currently ignored and reserved for future per-request tuning. Callers that
-// omit the key get a nil pointer so the router forwards "no preference" and
-// the websearch server falls back to its own default.
-func parseSafetyOptIn(raw any) *bool {
-	if raw == nil {
-		return nil
-	}
-	enabled := true
-	return &enabled
-}
-
-// safetyOptIns captures the caller's per-request opt-in for the websearch
-// server's PII and prompt-injection filters so `connectToolSession` can turn
-// them into `X-Tinfoil-Tool-*` headers on the MCP session. A nil pointer means
-// the caller said nothing and the server should apply its own default.
-type safetyOptIns struct {
-	pii       *bool
-	injection *bool
-}
-
-// parseSafetyOptIns pulls the caller-provided PII and prompt-injection opt-ins
-// off the top-level request body. Both Chat Completions and Responses expose
-// the controls as top-level keys (`pii_check_options`,
-// `prompt_injection_check_options`) so a single parser covers both routes.
-func parseSafetyOptIns(body map[string]any) safetyOptIns {
-	return safetyOptIns{
-		pii:       parseSafetyOptIn(body["pii_check_options"]),
-		injection: parseSafetyOptIn(body["prompt_injection_check_options"]),
-	}
-}
-
-// extractUserLocationCountry pulls the ISO 3166-1 alpha-2 country code out of
-// OpenAI's user_location object, which nests the actual country under an
-// `approximate` sub-object.
-func extractUserLocationCountry(raw any) string {
-	loc, _ := raw.(map[string]any)
-	if len(loc) == 0 {
-		return ""
-	}
-	if approx, ok := loc["approximate"].(map[string]any); ok {
-		if country := strings.TrimSpace(stringValue(approx["country"])); country != "" {
-			return strings.ToUpper(country)
-		}
-	}
-	return strings.ToUpper(strings.TrimSpace(stringValue(loc["country"])))
-}
-
-// extractAllowedDomains reads filters.allowed_domains from either a
-// web_search_options block or a Responses tool entry.
-func extractAllowedDomains(raw any) []string {
-	filters, ok := raw.(map[string]any)
-	if !ok {
-		return nil
-	}
-	return extractDomainListFromFilterMap(filters, "allowed_domains")
-}
-
-// extractExcludedDomains mirrors extractAllowedDomains for the
-// filters.excluded_domains list used to drop hosts from search results.
-func extractExcludedDomains(raw any) []string {
-	filters, ok := raw.(map[string]any)
-	if !ok {
-		return nil
-	}
-	return extractDomainListFromFilterMap(filters, "excluded_domains")
-}
-
-// extractAllowedDomainsFromFilterMap is kept as a thin shim over the shared
-// domain-list helper so existing callers keep compiling; new code should call
-// extractDomainListFromFilterMap directly.
-func extractAllowedDomainsFromFilterMap(filters map[string]any) []string {
-	return extractDomainListFromFilterMap(filters, "allowed_domains")
-}
-
-// extractDomainListFromFilterMap reads an array of hostnames under `key` from
-// a filters map and returns it lowercased, trimmed, and deduplicated.
-func extractDomainListFromFilterMap(filters map[string]any, key string) []string {
-	raw, ok := filters[key].([]any)
-	if !ok {
-		return nil
-	}
-	out := make([]string, 0, len(raw))
-	seen := make(map[string]struct{}, len(raw))
-	for _, item := range raw {
-		domain := strings.ToLower(strings.TrimSpace(stringValue(item)))
-		if domain == "" {
-			continue
-		}
-		if _, dup := seen[domain]; dup {
-			continue
-		}
-		seen[domain] = struct{}{}
-		out = append(out, domain)
-	}
-	return out
-}
-
-// applyAdvancedSearchFields lifts the non-domain search knobs (content_mode,
-// max_content_chars, category, publish date range, max_age_hours) off a
-// caller-provided block into opts. Callers are responsible for the
-// surrounding domain/country parsing so this stays focused on the fields that
-// are identical across Chat and Responses surfaces.
-func applyAdvancedSearchFields(opts *webSearchOptions, raw map[string]any) {
-	if opts == nil || raw == nil {
-		return
-	}
-	if mode := strings.TrimSpace(stringValue(raw["content_mode"])); mode != "" {
-		opts.contentMode = mode
-	}
-	if n, ok := intValue(raw["max_content_chars"]); ok && n > 0 {
-		opts.maxContentChars = n
-	}
-	if cat := strings.TrimSpace(stringValue(raw["category"])); cat != "" {
-		opts.category = cat
-	}
-	if start := normalizePublishedDate(raw["start_published_date"]); start != "" {
-		opts.startPublishedDate = start
-	}
-	if end := normalizePublishedDate(raw["end_published_date"]); end != "" {
-		opts.endPublishedDate = end
-	}
-	if n, ok := intValue(raw["max_age_hours"]); ok {
-		v := n
-		opts.maxAgeHours = &v
-	}
-}
-
-// intValue coerces a JSON-decoded number or numeric string into an int. The
-// second return reports whether the input was a recognizable integer; callers
-// can use that to distinguish "unset" from "explicitly zero", which matters
-// for tri-state fields like max_age_hours.
-func intValue(raw any) (int, bool) {
-	switch v := raw.(type) {
-	case nil:
-		return 0, false
-	case int:
-		return v, true
-	case int64:
-		return int(v), true
-	case float64:
-		return int(v), true
-	case json.Number:
-		if n, err := v.Int64(); err == nil {
-			return int(n), true
-		}
-	case string:
-		trimmed := strings.TrimSpace(v)
-		if trimmed == "" {
-			return 0, false
-		}
-		if n, err := strconv.Atoi(trimmed); err == nil {
-			return n, true
-		}
-	}
-	return 0, false
-}
-
-// normalizePublishedDate accepts `YYYY-MM-DD` or RFC 3339 strings and returns
-// the value as RFC 3339 in UTC. Empty or malformed input is dropped rather
-// than surfaced as an error so a single bad date does not fail the whole
-// request; the MCP server still gets a syntactically valid value or nothing.
-func normalizePublishedDate(raw any) string {
-	str := strings.TrimSpace(stringValue(raw))
-	if str == "" {
-		return ""
-	}
-	if t, err := time.Parse("2006-01-02", str); err == nil {
-		return t.UTC().Format(time.RFC3339)
-	}
-	if t, err := time.Parse(time.RFC3339, str); err == nil {
-		return t.UTC().Format(time.RFC3339)
-	}
-	return ""
-}
-
-// applyWebSearchOptionsToToolCall dispatches to the per-tool merge helper so
-// options forwarded from OpenAI's request shape end up on the MCP tool call.
-func applyWebSearchOptionsToToolCall(toolName string, arguments map[string]any, opts webSearchOptions) {
-	switch toolName {
-	case "search":
-		opts.applyToSearchArgs(arguments)
-	case "fetch":
-		opts.applyToFetchArgs(arguments)
-	}
-}
-
-// toAnySlice lifts a []string into a []any so it marshals cleanly through the
-// generic map[string]any request body the MCP client expects.
-func toAnySlice(values []string) []any {
-	out := make([]any, len(values))
-	for i, value := range values {
-		out[i] = value
-	}
-	return out
 }
 
 type headerRoundTripper struct {
@@ -1059,1278 +556,7 @@ func postJSON(ctx context.Context, em *manager.EnclaveManager, modelName, path s
 	}, nil
 }
 
-type toolCall struct {
-	id        string
-	name      string
-	arguments map[string]any
-}
 
-func parseChatToolCalls(response map[string]any) (map[string]any, []toolCall) {
-	choices, _ := response["choices"].([]any)
-	if len(choices) == 0 {
-		return nil, nil
-	}
-	choice, _ := choices[0].(map[string]any)
-	message, _ := choice["message"].(map[string]any)
-	if message == nil {
-		return nil, nil
-	}
-	rawCalls, _ := message["tool_calls"].([]any)
-	toolCalls := make([]toolCall, 0, len(rawCalls))
-	for _, rawCall := range rawCalls {
-		callMap, _ := rawCall.(map[string]any)
-		functionMap, _ := callMap["function"].(map[string]any)
-		args := parseArguments(functionMap["arguments"])
-		toolCalls = append(toolCalls, toolCall{
-			id:        stringValue(callMap["id"]),
-			name:      stringValue(functionMap["name"]),
-			arguments: args,
-		})
-	}
-	return message, toolCalls
-}
-
-func parseResponsesToolCalls(response map[string]any) []toolCall {
-	output, _ := response["output"].([]any)
-	var result []toolCall
-	for _, item := range output {
-		itemMap, _ := item.(map[string]any)
-		itemType := stringValue(itemMap["type"])
-		if itemType != "function_call" && itemType != "mcp_call" {
-			continue
-		}
-		callID := stringValue(itemMap["call_id"])
-		if callID == "" {
-			callID = stringValue(itemMap["id"])
-		}
-		name := stringValue(itemMap["name"])
-		args := parseArguments(itemMap["arguments"])
-		result = append(result, toolCall{
-			id:        callID,
-			name:      name,
-			arguments: args,
-		})
-	}
-	return result
-}
-
-func parseArguments(raw any) map[string]any {
-	switch value := raw.(type) {
-	case string:
-		var parsed map[string]any
-		raw, _ := sanitizeToolCallArgumentsJSON([]byte(value))
-		if json.Unmarshal(raw, &parsed) == nil {
-			return parsed
-		}
-	case map[string]any:
-		return value
-	}
-	return map[string]any{}
-}
-
-type citationSource struct {
-	index int
-	url   string
-	title string
-}
-
-// toolCallRecord captures a tool call the router made on the user's behalf,
-// used to surface web_search_call progress items to clients. errorReason
-// carries the tool-side error message when the call failed so terminal
-// web_search_call items can honestly report status:"failed" instead of
-// silently claiming completion on a request that never returned results.
-type toolCallRecord struct {
-	name        string
-	arguments   map[string]any
-	resultURLs  []string
-	errorReason string
-}
-
-type citationState struct {
-	nextIndex int
-	sources   []citationSource
-	toolCalls []toolCallRecord
-}
-
-// recordToolCall appends a completed tool call (search or fetch) for later
-// surfacing as a web_search_call stream event or Responses output item.
-func (c *citationState) recordToolCall(record toolCallRecord) {
-	if c == nil {
-		return
-	}
-	c.toolCalls = append(c.toolCalls, record)
-}
-
-// record registers a source the router surfaced to the model in a tool output
-// so later passes over the model's final content can recognize inline markdown
-// links pointing at that URL and emit url_citation annotations for them.
-func (c *citationState) record(url, title string) int {
-	if c == nil {
-		return 1
-	}
-	if c.nextIndex <= 0 {
-		c.nextIndex = 1
-	}
-	index := c.nextIndex
-	c.nextIndex++
-	c.sources = append(c.sources, citationSource{index: index, url: url, title: title})
-	return index
-}
-
-// markdownLinkPattern matches inline markdown links of the form
-// `[visible text](https://example.com)` so the router can map the rendered
-// link span back to a recorded source URL. The label capture allows any
-// characters except a closing bracket, matching the subset of markdown the
-// model is asked to produce; nested brackets and images are not supported on
-// purpose.
-var markdownLinkPattern = regexp.MustCompile(`\[([^\]]+)\]\((https?://[^\s)]+)\)`)
-
-// fullwidthBracketedLinkPattern matches the near-markdown shape that some
-// web-search tuned models emit when they slip from ASCII brackets into
-// fullwidth lenticular brackets: `【visible text】(https://example.com)`.
-// It also accepts the mixed-bracket variant `【visible text](https://...)`
-// observed from gpt-oss when it improvises a citation outside the harmony
-// browser tool training distribution. The router rewrites matches to
-// canonical ASCII markdown before computing citation spans so every
-// downstream renderer sees the documented shape.
-var fullwidthBracketedLinkPattern = regexp.MustCompile(`\x{3010}([^\x{3011}\]]+)[\x{3011}\]]\((https?://[^\s)]+)\)`)
-
-// normalizeCitationLinks rewrites `【label】(url)` (and the `【label](url)`
-// mixed-bracket shape) occurrences in the model's final content into ASCII
-// markdown `[label](url)` so every consumer renders a clickable link
-// regardless of which bracket style the model produced.
-func normalizeCitationLinks(text string) string {
-	if text == "" || !strings.ContainsRune(text, '\u3010') {
-		return text
-	}
-	return fullwidthBracketedLinkPattern.ReplaceAllString(text, "[$1]($2)")
-}
-
-var toolOutputURLPattern = regexp.MustCompile(`(?m)^URL:\s*(\S+)`)
-
-// extractToolOutputURLs pulls the `URL: ...` lines the router embeds into each
-// numbered source block so the downstream stream emitter can surface them as
-// web_search_call `sources`.
-func extractToolOutputURLs(output string) []string {
-	matches := toolOutputURLPattern.FindAllStringSubmatch(output, -1)
-	if len(matches) == 0 {
-		return nil
-	}
-	urls := make([]string, 0, len(matches))
-	seen := make(map[string]struct{}, len(matches))
-	for _, match := range matches {
-		if len(match) < 2 {
-			continue
-		}
-		url := strings.TrimSpace(match[1])
-		if url == "" {
-			continue
-		}
-		if _, dup := seen[url]; dup {
-			continue
-		}
-		seen[url] = struct{}{}
-		urls = append(urls, url)
-	}
-	return urls
-}
-
-// toActionSources wraps a list of URLs in the OpenAI-spec shape
-// `[{type: "url", url: "..."}]` documented for `WebSearchCall.action.sources`.
-// Returns nil when the input is empty so callers can omit the field entirely.
-func toActionSources(urls []string) []any {
-	if len(urls) == 0 {
-		return nil
-	}
-	sources := make([]any, 0, len(urls))
-	for _, url := range urls {
-		if url == "" {
-			continue
-		}
-		sources = append(sources, map[string]any{
-			"type": "url",
-			"url":  url,
-		})
-	}
-	if len(sources) == 0 {
-		return nil
-	}
-	return sources
-}
-
-type annotationMatch struct {
-	// startIndex and endIndex span the visible label of the markdown link
-	// (i.e. the text between the square brackets), matching the shape of
-	// OpenAI's url_citation start_index/end_index fields.
-	startIndex int
-	endIndex   int
-	source     citationSource
-}
-
-// matchesFor scans the model's final content for inline markdown links whose
-// URL matches a source the router recorded during tool execution and returns
-// one annotationMatch per such occurrence. The start/end indices span the
-// link's visible label (the characters between the square brackets), which is
-// the convention OpenAI's web_search_call annotations use.
-//
-// Links that point at URLs the router never recorded are ignored so a model
-// that fabricates a URL cannot cause a broken annotation to ship. A given
-// label+URL pair can appear multiple times in the text; each occurrence
-// produces its own annotation to match how OpenAI's Responses API reports
-// repeated citations.
-func (c *citationState) matchesFor(text string) []annotationMatch {
-	if c == nil || len(c.sources) == 0 || text == "" {
-		return nil
-	}
-
-	byURL := make(map[string]citationSource, len(c.sources))
-	for _, source := range c.sources {
-		if source.url == "" {
-			continue
-		}
-		if existing, ok := byURL[source.url]; !ok || (existing.title == "" && source.title != "") {
-			byURL[source.url] = source
-		}
-	}
-	if len(byURL) == 0 {
-		return nil
-	}
-
-	raw := markdownLinkPattern.FindAllStringSubmatchIndex(text, -1)
-	if len(raw) == 0 {
-		return nil
-	}
-
-	byteToRune := newByteToRuneIndex(text)
-	matches := make([]annotationMatch, 0, len(raw))
-	for _, m := range raw {
-		if len(m) < 6 {
-			continue
-		}
-		url := text[m[4]:m[5]]
-		source, ok := byURL[url]
-		if !ok {
-			continue
-		}
-		// Span the visible label, not the whole [label](url) expression, so
-		// downstream consumers can highlight just the link text the user sees.
-		// Convert regex byte offsets to Unicode code-point offsets so the
-		// indices match what OpenAI SDKs and JS/Python clients observe when
-		// they index strings by character.
-		matches = append(matches, annotationMatch{
-			startIndex: byteToRune.convert(m[2]),
-			endIndex:   byteToRune.convert(m[3]),
-			source:     source,
-		})
-	}
-	return matches
-}
-
-// byteToRuneIndex translates byte offsets into rune (code point) offsets for
-// a given string. It caches the last lookup so sequential calls, as produced
-// by a left-to-right scan of regex matches, stay linear in the source length.
-type byteToRuneIndex struct {
-	text     string
-	lastByte int
-	lastRune int
-}
-
-func newByteToRuneIndex(text string) *byteToRuneIndex {
-	return &byteToRuneIndex{text: text}
-}
-
-func (b *byteToRuneIndex) convert(byteOffset int) int {
-	if byteOffset <= b.lastByte {
-		b.lastByte = 0
-		b.lastRune = 0
-	}
-	for b.lastByte < byteOffset && b.lastByte < len(b.text) {
-		_, size := utf8.DecodeRuneInString(b.text[b.lastByte:])
-		if size == 0 {
-			break
-		}
-		b.lastByte += size
-		b.lastRune++
-	}
-	return b.lastRune
-}
-
-// nestedAnnotationsFor returns annotations in the Chat Completions API shape
-// documented by OpenAI:
-//
-//	{"type":"url_citation","url_citation":{"title":..,"url":..,"start_index":..,"end_index":..}}
-func (c *citationState) nestedAnnotationsFor(text string) []any {
-	matches := c.matchesFor(text)
-	if len(matches) == 0 {
-		return nil
-	}
-	annotations := make([]any, 0, len(matches))
-	for _, match := range matches {
-		citation := map[string]any{
-			"url":         match.source.url,
-			"start_index": match.startIndex,
-			"end_index":   match.endIndex,
-		}
-		if match.source.title != "" {
-			citation["title"] = match.source.title
-		}
-		annotations = append(annotations, map[string]any{
-			"type":         "url_citation",
-			"url_citation": citation,
-		})
-	}
-	return annotations
-}
-
-// flatAnnotationsFor returns annotations in the Responses API shape documented
-// by OpenAI, where url_citation fields live directly on the annotation object:
-//
-//	{"type":"url_citation","start_index":..,"end_index":..,"url":..,"title":..}
-func (c *citationState) flatAnnotationsFor(text string) []any {
-	matches := c.matchesFor(text)
-	if len(matches) == 0 {
-		return nil
-	}
-	annotations := make([]any, 0, len(matches))
-	for _, match := range matches {
-		annotation := map[string]any{
-			"type":        "url_citation",
-			"start_index": match.startIndex,
-			"end_index":   match.endIndex,
-			"url":         match.source.url,
-		}
-		if match.source.title != "" {
-			annotation["title"] = match.source.title
-		}
-		annotations = append(annotations, annotation)
-	}
-	return annotations
-}
-
-// publicToolErrorReason returns a short, opaque status string safe to
-// ship to clients via `web_search_call.reason`. The raw error text is
-// recorded to the server log so operators can still diagnose failures
-// without having to surface internal hostnames, gRPC error bodies, or
-// other implementation details to end users.
-//
-// Safety-blocked errors (PII or prompt-injection safeguards tripping on the
-// caller's query) return the distinct `blockedToolErrorReason` so the
-// caller can render them with the dedicated `blocked` web_search_call
-// status instead of collapsing them into a generic failure.
-const (
-	publicToolErrorReasonString = "tool_error"
-	blockedToolErrorReason      = "blocked_by_safety_filter"
-)
-
-func publicToolErrorReason(toolName string, err error) string {
-	if err == nil {
-		return ""
-	}
-	log.Printf("toolruntime: %s tool call failed: %v", toolName, err)
-	if isToolCallBlocked(err) {
-		return blockedToolErrorReason
-	}
-	return publicToolErrorReasonString
-}
-
-// isToolCallBlocked reports whether the MCP tool error came from a PII or
-// prompt-injection safeguard. Safety-blocked tool errors are wrapped with
-// a message beginning `query was blocked by safety filters`. Detecting
-// that prefix lets the router surface `status: "blocked"` on the
-// corresponding web_search_call envelope so clients can render a distinct
-// affordance without exposing the raw detail string.
-func isToolCallBlocked(err error) bool {
-	if err == nil {
-		return false
-	}
-	return strings.Contains(strings.ToLower(err.Error()), "blocked by safety filter")
-}
-
-// failureStatusFor picks the web_search_call status string to surface for
-// a non-nil MCP error. Safety-blocked errors become `"blocked"` so clients
-// can distinguish them from generic tool failures; everything else stays
-// `"failed"`.
-func failureStatusFor(err error) string {
-	if isToolCallBlocked(err) {
-		return "blocked"
-	}
-	return "failed"
-}
-
-// toolResultErrorMessage extracts a human-readable error message from a
-// CallToolResult whose `IsError` flag is set. The MCP SDK packs handler
-// errors into the result's text content rather than returning them as
-// protocol errors, so callers need this helper to surface the underlying
-// message (e.g. "query was blocked by safety filters: ...") to the rest
-// of the router for classification and user-visible reporting.
-func toolResultErrorMessage(result *mcp.CallToolResult) string {
-	if result == nil {
-		return ""
-	}
-	var parts []string
-	for _, content := range result.Content {
-		if textContent, ok := content.(*mcp.TextContent); ok {
-			if trimmed := strings.TrimSpace(textContent.Text); trimmed != "" {
-				parts = append(parts, trimmed)
-			}
-		}
-	}
-	return strings.Join(parts, "\n")
-}
-
-func callTool(ctx context.Context, session *mcp.ClientSession, name string, arguments map[string]any, citations *citationState) (string, error) {
-	start := time.Now()
-	result, err := session.CallTool(ctx, &mcp.CallToolParams{
-		Name:      name,
-		Arguments: arguments,
-	})
-	if err != nil {
-		debugLogf("toolruntime:mcp.error tool=%s elapsed=%s args=%s err=%v", name, time.Since(start), debugPreview(arguments, 400), err)
-		return "", err
-	}
-	if result.IsError {
-		message := toolResultErrorMessage(result)
-		debugLogf("toolruntime:mcp.tool_error tool=%s elapsed=%s args=%s message=%q", name, time.Since(start), debugPreview(arguments, 400), message)
-		if message == "" {
-			message = "tool call failed"
-		}
-		return "", errors.New(message)
-	}
-	if debugEnabled {
-		hasStructured := result.StructuredContent != nil
-		textParts := 0
-		for _, content := range result.Content {
-			if _, ok := content.(*mcp.TextContent); ok {
-				textParts++
-			}
-		}
-		debugLogf("toolruntime:mcp.ok tool=%s elapsed=%s structured=%t text_parts=%d structured_preview=%s",
-			name, time.Since(start), hasStructured, textParts, debugPreview(result.StructuredContent, 500))
-	}
-	if result.StructuredContent != nil {
-		if formatted := formatStructuredToolOutput(name, result.StructuredContent, citations); formatted != "" {
-			return formatted, nil
-		}
-		body, err := json.Marshal(result.StructuredContent)
-		if err == nil {
-			return string(body), nil
-		}
-	}
-	var parts []string
-	for _, content := range result.Content {
-		if textContent, ok := content.(*mcp.TextContent); ok {
-			parts = append(parts, textContent.Text)
-		}
-	}
-	return strings.Join(parts, "\n"), nil
-}
-
-func formatStructuredToolOutput(name string, raw any, citations *citationState) string {
-	content, _ := raw.(map[string]any)
-	if len(content) == 0 {
-		return ""
-	}
-
-	switch name {
-	case "search":
-		return formatSearchToolOutput(content["results"], citations)
-	case "fetch":
-		if formatted := formatFetchToolOutput(content["pages"], citations); formatted != "" {
-			return formatted
-		}
-		return formatFetchFailures(content["results"])
-	default:
-		return ""
-	}
-}
-
-func formatSearchToolOutput(raw any, citations *citationState) string {
-	results, _ := raw.([]any)
-	if len(results) == 0 {
-		return "No safe search results were found."
-	}
-
-	var out strings.Builder
-	for _, rawResult := range results {
-		result, _ := rawResult.(map[string]any)
-		if result == nil {
-			continue
-		}
-
-		title := strings.TrimSpace(stringValue(result["title"]))
-		url := strings.TrimSpace(stringValue(result["url"]))
-		content := strings.TrimSpace(stringValue(result["content"]))
-		published := strings.TrimSpace(stringValue(result["published_date"]))
-		citations.record(url, title)
-
-		if title == "" {
-			title = "Search result"
-		}
-		fmt.Fprintf(&out, "Source: %s\n", title)
-		if url != "" {
-			fmt.Fprintf(&out, "URL: %s\n", url)
-		}
-		if published != "" {
-			fmt.Fprintf(&out, "Published: %s\n", published)
-		}
-		if content != "" {
-			out.WriteString(content)
-			out.WriteString("\n")
-		}
-		out.WriteString("\n")
-	}
-	return strings.TrimSpace(out.String())
-}
-
-func formatFetchToolOutput(raw any, citations *citationState) string {
-	pages, _ := raw.([]any)
-	if len(pages) == 0 {
-		return ""
-	}
-
-	var out strings.Builder
-	for _, rawPage := range pages {
-		page, _ := rawPage.(map[string]any)
-		if page == nil {
-			continue
-		}
-
-		url := strings.TrimSpace(stringValue(page["url"]))
-		content := strings.TrimSpace(stringValue(page["content"]))
-		citations.record(url, "Fetched page")
-		out.WriteString("Source: Fetched page\n")
-		if url != "" {
-			fmt.Fprintf(&out, "URL: %s\n", url)
-		}
-		if content != "" {
-			out.WriteString(content)
-			out.WriteString("\n")
-		}
-		out.WriteString("\n")
-	}
-	return strings.TrimSpace(out.String())
-}
-
-func formatFetchFailures(raw any) string {
-	results, _ := raw.([]any)
-	if len(results) == 0 {
-		return ""
-	}
-
-	lines := make([]string, 0, len(results))
-	for _, rawResult := range results {
-		result, _ := rawResult.(map[string]any)
-		if result == nil {
-			continue
-		}
-		status := strings.TrimSpace(stringValue(result["status"]))
-		if status == "completed" {
-			continue
-		}
-		url := strings.TrimSpace(stringValue(result["url"]))
-		errText := strings.TrimSpace(stringValue(result["error"]))
-		switch {
-		case url != "" && errText != "":
-			lines = append(lines, fmt.Sprintf("Fetch failed for %s: %s", url, errText))
-		case url != "":
-			lines = append(lines, fmt.Sprintf("Fetch failed for %s.", url))
-		case errText != "":
-			lines = append(lines, fmt.Sprintf("Fetch failed: %s", errText))
-		}
-	}
-	return strings.Join(lines, "\n")
-}
-
-// attachChatCitations resolves the inline markdown links the model wrote into
-// each choice's content back to the URLs the router registered during the
-// tool loop and attaches them to message.annotations in the Chat Completions
-// nested url_citation shape:
-//
-//	{"type":"url_citation","url_citation":{"title":..,"url":..,"start_index":..,"end_index":..}}
-//
-// This matches OpenAI's documented Chat Completions response shape.
-//
-// When `eventsEnabled` is true, `<tinfoil-event>...</tinfoil-event>`
-// markers for every recorded router tool call are prepended to each
-// choice's message content so clients that opted into the tinfoil-event
-// stream see the same in_progress -> terminal progression in the
-// non-streaming response as they do in streaming. url_citation
-// annotation indices are shifted by the marker prefix length so spans
-// continue to refer to the correct byte range in the new content.
-func attachChatCitations(responseBody map[string]any, citations *citationState, eventsEnabled bool) {
-	if responseBody == nil || citations == nil {
-		return
-	}
-	choices, _ := responseBody["choices"].([]any)
-	for _, rawChoice := range choices {
-		choice, _ := rawChoice.(map[string]any)
-		if choice == nil {
-			continue
-		}
-		message, _ := choice["message"].(map[string]any)
-		if message == nil {
-			continue
-		}
-		content := stringValue(message["content"])
-		if normalized := normalizeCitationLinks(content); normalized != content {
-			content = normalized
-		}
-		annotations := citations.nestedAnnotationsFor(content)
-		var prefix string
-		if eventsEnabled {
-			prefix = tinfoilEventMarkersForRecords(citations.toolCalls)
-		}
-		if prefix != "" {
-			content = prefix + content
-			shiftNestedAnnotationIndices(annotations, utf8.RuneCountInString(prefix))
-		}
-		message["content"] = content
-		if len(annotations) > 0 {
-			message["annotations"] = annotations
-		}
-	}
-}
-
-// shiftNestedAnnotationIndices shifts every `start_index` / `end_index`
-// on a nested-shape url_citation by `offset` bytes. Called after marker
-// injection to keep the annotation spans pointing at the same substrings
-// of the (now longer) content string.
-func shiftNestedAnnotationIndices(annotations []any, offset int) {
-	if offset == 0 {
-		return
-	}
-	for _, raw := range annotations {
-		ann, _ := raw.(map[string]any)
-		if ann == nil {
-			continue
-		}
-		inner, _ := ann["url_citation"].(map[string]any)
-		if inner == nil {
-			continue
-		}
-		if start, ok := inner["start_index"].(int); ok {
-			inner["start_index"] = start + offset
-		}
-		if end, ok := inner["end_index"].(int); ok {
-			inner["end_index"] = end + offset
-		}
-	}
-}
-
-// shiftFlatAnnotationIndices is the Responses-shape counterpart of
-// shiftNestedAnnotationIndices: it mutates every `start_index` /
-// `end_index` on a flat url_citation annotation list so the spans
-// continue to reference the same substrings after a marker prefix is
-// injected into the output_text content.
-func shiftFlatAnnotationIndices(annotations []any, offset int) {
-	if offset == 0 {
-		return
-	}
-	for _, raw := range annotations {
-		ann, _ := raw.(map[string]any)
-		if ann == nil {
-			continue
-		}
-		if start, ok := ann["start_index"].(int); ok {
-			ann["start_index"] = start + offset
-		}
-		if end, ok := ann["end_index"].(int); ok {
-			ann["end_index"] = end + offset
-		}
-	}
-}
-
-// attachResponsesCitations resolves inline markdown links in each output_text
-// item back to the URLs the router registered during the tool loop and
-// attaches them in the Responses API flat url_citation shape documented by
-// OpenAI:
-//
-//	{"type":"url_citation","start_index":..,"end_index":..,"url":..,"title":..}
-//
-// It also prepends a web_search_call output item per recorded router tool
-// call in the shape OpenAI documents for the Responses API, surfaced to
-// clients as response.output_item.added events.
-//
-// When `eventsEnabled` is true, `<tinfoil-event>` progress markers for
-// every recorded router tool call are prepended to the first assistant
-// output_text content block so non-streaming clients that opted into the
-// marker stream see the same in_progress -> terminal progression they
-// get over the streaming carrier. url_citation `start_index` /
-// `end_index` are shifted by the marker prefix length so spans keep
-// pointing at the correct bytes in the new content.
-func attachResponsesCitations(responseBody map[string]any, citations *citationState, includeActionSources, eventsEnabled bool) {
-	if responseBody == nil || citations == nil {
-		return
-	}
-	outputItems, _ := responseBody["output"].([]any)
-	markerPrefix := ""
-	if eventsEnabled {
-		markerPrefix = tinfoilEventMarkersForRecords(citations.toolCalls)
-	}
-	markerInjected := false
-	for _, rawItem := range outputItems {
-		item, _ := rawItem.(map[string]any)
-		if item == nil || stringValue(item["type"]) != "message" {
-			continue
-		}
-		contentList, _ := item["content"].([]any)
-		for _, rawContent := range contentList {
-			contentMap, _ := rawContent.(map[string]any)
-			if contentMap == nil || stringValue(contentMap["type"]) != "output_text" {
-				continue
-			}
-			text := stringValue(contentMap["text"])
-			if normalized := normalizeCitationLinks(text); normalized != text {
-				contentMap["text"] = normalized
-				text = normalized
-			}
-			annotations := citations.flatAnnotationsFor(text)
-			if !markerInjected && markerPrefix != "" {
-				text = markerPrefix + text
-				contentMap["text"] = text
-				shiftFlatAnnotationIndices(annotations, utf8.RuneCountInString(markerPrefix))
-				markerInjected = true
-			}
-			if len(annotations) > 0 {
-				contentMap["annotations"] = annotations
-			}
-		}
-	}
-
-	// If events were requested but no assistant message carried an
-	// output_text to prefix onto, synthesize a minimal message item so
-	// the marker still reaches the client. This keeps parity with the
-	// chat non-streaming path, which always has a `content` string to
-	// prepend to.
-	if !markerInjected && markerPrefix != "" {
-		synthetic := map[string]any{
-			"id":     "msg_tf_" + uuid.NewString(),
-			"type":   "message",
-			"role":   "assistant",
-			"status": "completed",
-			"content": []any{
-				map[string]any{"type": "output_text", "text": markerPrefix, "annotations": []any{}},
-			},
-		}
-		outputItems = append([]any{synthetic}, outputItems...)
-		responseBody["output"] = outputItems
-	}
-
-	if len(citations.toolCalls) == 0 {
-		return
-	}
-	events := buildWebSearchCallOutputItems(citations.toolCalls, includeActionSources)
-	prepended := make([]any, 0, len(events)+len(outputItems))
-	prepended = append(prepended, events...)
-	prepended = append(prepended, outputItems...)
-	responseBody["output"] = prepended
-}
-
-// buildWebSearchCallOutputItems turns recorded router tool calls into the
-// web_search_call output items documented by OpenAI's Responses API.
-//   - search tool calls become one action.type:"search" event with the query
-//     and (when the caller opted in via `include:
-//     ["web_search_call.action.sources"]`) the resolved source URLs in the
-//     spec shape `[{type:"url", url:"..."}]`.
-//   - fetch tool calls become one action.type:"open_page" event per URL so
-//     clients can correlate each fetched page with its surrounding search.
-func buildWebSearchCallOutputItems(records []toolCallRecord, includeActionSources bool) []any {
-	events := make([]any, 0, len(records))
-	for _, record := range records {
-		status := statusForRecord(record)
-		switch record.name {
-		case "search":
-			action := map[string]any{"type": "search"}
-			if query := stringValue(record.arguments["query"]); query != "" {
-				action["query"] = query
-			}
-			if includeActionSources {
-				if sources := toActionSources(record.resultURLs); len(sources) > 0 {
-					action["sources"] = sources
-				}
-			}
-			events = append(events, webSearchCallEvent("ws_"+uuid.NewString(), status, action))
-		case "fetch":
-			for _, url := range fetchArgumentURLs(record.arguments) {
-				action := map[string]any{"type": "open_page", "url": url}
-				events = append(events, webSearchCallEvent("ws_"+uuid.NewString(), status, action))
-			}
-		}
-	}
-	return events
-}
-
-// statusForRecord maps a recorded router tool call to the web_search_call
-// status surfaced to clients. "blocked" is reserved for PII or prompt
-// injection safeguards so client UIs can surface a distinct affordance for
-// the user; any other error becomes "failed".
-func statusForRecord(record toolCallRecord) string {
-	if record.errorReason == "" {
-		return "completed"
-	}
-	if record.errorReason == blockedToolErrorReason {
-		return "blocked"
-	}
-	return "failed"
-}
-
-// webSearchCallEvent builds a single web_search_call output item in the
-// spec-conformant shape documented by OpenAI for the Responses API. The
-// returned map is used only for non-streaming `output[]` items; it does
-// NOT carry the `item_id` / `reason` fields the streaming envelopes use,
-// and it maps the router's internal `blocked` status onto the spec-valid
-// `failed` status (the safety-block signal is surfaced separately via
-// tinfoil-event markers when the caller opts in).
-func webSearchCallEvent(id, status string, action map[string]any) map[string]any {
-	if status == "blocked" {
-		status = "failed"
-	}
-	return map[string]any{
-		"type":   "web_search_call",
-		"id":     id,
-		"status": status,
-		"action": action,
-	}
-}
-
-// tinfoilEventsEnabled reports whether the caller opted into the
-// router-owned tinfoil-event marker stream for web-search tool calls.
-// The header value is a comma-separated list of families the client is
-// prepared to parse. We match case-insensitively and tolerate whitespace.
-// A missing or empty header means the client gets a pristine
-// spec-conformant stream with no markers at all.
-func tinfoilEventsEnabled(h http.Header) bool {
-	if h == nil {
-		return false
-	}
-	raw := h.Get(tinfoilEventsHeader)
-	if raw == "" {
-		return false
-	}
-	for _, part := range strings.Split(raw, ",") {
-		if strings.EqualFold(strings.TrimSpace(part), tinfoilEventsWebSearch) {
-			return true
-		}
-	}
-	return false
-}
-
-// tinfoilEventMarker renders a single progress event as a text marker:
-// a JSON payload wrapped in `<tinfoil-event>...</tinfoil-event>` tags on
-// its own line. The leading and trailing newlines isolate the marker
-// from natural-language text so client strip regexes do not leave stray
-// whitespace, and so raw SSE captures stay readable for debugging.
-func tinfoilEventMarker(id, status string, action map[string]any, reason string) string {
-	payload := map[string]any{
-		"type":    tinfoilEventPayloadType,
-		"item_id": id,
-		"status":  status,
-		"action":  action,
-	}
-	if reason != "" {
-		payload["error"] = map[string]any{"code": reason}
-	}
-	data, err := json.Marshal(payload)
-	if err != nil {
-		// json.Marshal on a map with string keys and primitive /
-		// map[string]any values cannot fail in practice. Fall back to
-		// an empty payload rather than panicking so a bug here can
-		// never break the main stream for a client.
-		data = []byte(`{"type":"` + tinfoilEventPayloadType + `"}`)
-	}
-	return "\n" + tinfoilEventOpenTag + string(data) + tinfoilEventCloseTag + "\n"
-}
-
-// tinfoilEventMarkersForRecords renders a sequence of in_progress then
-// terminal markers (completed / failed / blocked) for the non-streaming
-// paths, which only see recorded tool calls after they have already run.
-// Each recorded `search` call yields one marker pair; each recorded
-// `fetch` call yields one marker pair per URL so clients see the same
-// progression as the streaming paths. The non-streaming carrier keeps
-// the distinct `blocked` status inside the marker JSON even though the
-// spec-conformant `web_search_call` output item collapses it onto
-// `failed` — the whole point of the marker is to surface details the
-// spec has no slot for.
-func tinfoilEventMarkersForRecords(records []toolCallRecord) string {
-	if len(records) == 0 {
-		return ""
-	}
-	var builder strings.Builder
-	status := func(record toolCallRecord) string { return statusForRecord(record) }
-	writePair := func(action map[string]any, s, reason string) {
-		id := "ws_" + uuid.NewString()
-		builder.WriteString(tinfoilEventMarker(id, "in_progress", action, ""))
-		builder.WriteString(tinfoilEventMarker(id, s, action, reason))
-	}
-	for _, record := range records {
-		switch record.name {
-		case "search":
-			action := map[string]any{"type": "search"}
-			if query := stringValue(record.arguments["query"]); query != "" {
-				action["query"] = query
-			}
-			writePair(action, status(record), record.errorReason)
-		case "fetch":
-			urls := fetchArgumentURLs(record.arguments)
-			if len(urls) == 0 {
-				continue
-			}
-			for _, url := range urls {
-				action := map[string]any{"type": "open_page", "url": url}
-				writePair(action, status(record), record.errorReason)
-			}
-		}
-	}
-	return builder.String()
-}
-
-// fetchArgumentURLs extracts the string URLs the model handed the fetch tool.
-func fetchArgumentURLs(arguments map[string]any) []string {
-	raw, ok := arguments["urls"].([]any)
-	if !ok {
-		return nil
-	}
-	urls := make([]string, 0, len(raw))
-	for _, item := range raw {
-		if url := strings.TrimSpace(stringValue(item)); url != "" {
-			urls = append(urls, url)
-		}
-	}
-	return urls
-}
-
-func ownedToolNames(tools []*mcp.Tool) map[string]struct{} {
-	result := make(map[string]struct{}, len(tools))
-	for _, tool := range tools {
-		if tool == nil || tool.Name == "" {
-			continue
-		}
-		result[tool.Name] = struct{}{}
-	}
-	return result
-}
-
-func splitToolCalls(ownedTools map[string]struct{}, toolCalls []toolCall) ([]toolCall, []toolCall) {
-	routerToolCalls := make([]toolCall, 0, len(toolCalls))
-	clientToolCalls := make([]toolCall, 0, len(toolCalls))
-	for _, toolCall := range toolCalls {
-		if _, ok := ownedTools[toolCall.name]; ok {
-			routerToolCalls = append(routerToolCalls, toolCall)
-			continue
-		}
-		clientToolCalls = append(clientToolCalls, toolCall)
-	}
-	return routerToolCalls, clientToolCalls
-}
-
-func existingTools(raw any) []any {
-	tools, _ := raw.([]any)
-	if tools == nil {
-		return []any{}
-	}
-	return tools
-}
-
-func chatTools(tools []*mcp.Tool) []any {
-	result := make([]any, 0, len(tools))
-	for _, tool := range tools {
-		result = append(result, map[string]any{
-			"type": "function",
-			"function": map[string]any{
-				"name":        tool.Name,
-				"description": routedToolDescription(tool),
-				"parameters":  tool.InputSchema,
-			},
-		})
-	}
-	return result
-}
-
-func responseTools(tools []*mcp.Tool) []any {
-	result := make([]any, 0, len(tools))
-	for _, tool := range tools {
-		result = append(result, map[string]any{
-			"type":        "function",
-			"name":        tool.Name,
-			"description": routedToolDescription(tool),
-			"parameters":  tool.InputSchema,
-		})
-	}
-	return result
-}
-
-func replaceResponsesWebSearchTools(raw any, replacements []any) []any {
-	tools, _ := raw.([]any)
-	if len(tools) == 0 {
-		return replacements
-	}
-	result := make([]any, 0, len(tools)+len(replacements))
-	injected := false
-	for _, tool := range tools {
-		toolMap, _ := tool.(map[string]any)
-		if stringValue(toolMap["type"]) == "web_search" {
-			if !injected {
-				result = append(result, replacements...)
-				injected = true
-			}
-			continue
-		}
-		result = append(result, tool)
-	}
-	return result
-}
-
-func prependChatPrompt(prompt *mcp.GetPromptResult, raw any) []any {
-	messages, _ := raw.([]any)
-	prefix := chatPromptPrefix(prompt)
-	if len(prefix) == 0 {
-		return messages
-	}
-	return append(prefix, messages...)
-}
-
-func prependResponsesPrompt(prompt *mcp.GetPromptResult, raw any) any {
-	items := normalizeResponsesInput(raw)
-	prefix := responsePromptPrefix(prompt)
-	if len(prefix) == 0 {
-		return items
-	}
-	return append(prefix, items...)
-}
-
-func buildRouterPrompt() *mcp.GetPromptResult {
-	return &mcp.GetPromptResult{
-		Description: "Instructions for router-owned web search tool use.",
-		Messages: []*mcp.PromptMessage{
-			{
-				Role: "system",
-				Content: &mcp.TextContent{
-					Text: "You may use the search and fetch tools when current web information would improve the answer. Use search first to discover sources, then fetch specific URLs only when you need deeper detail. " + citationInstructions + " " + toolOutputWarning + " " + toolEconomyInstructions,
-				},
-			},
-		},
-	}
-}
-
-func forcedFinalChatRequest(reqBody map[string]any) map[string]any {
-	finalBody := cloneJSONMap(reqBody)
-	messages, _ := finalBody["messages"].([]any)
-	finalBody["messages"] = append(messages, map[string]any{
-		"role":    "system",
-		"content": finalAnswerInstructionText,
-	})
-	delete(finalBody, "tools")
-	// Every field that could otherwise require or steer a tool call must
-	// be stripped too; leaving tool_choice:"required" on a request that
-	// carries no tools is a protocol-level contradiction that upstream
-	// will reject or loop on. function_call is the legacy Chat
-	// Completions equivalent and is removed for the same reason.
-	delete(finalBody, "tool_choice")
-	delete(finalBody, "function_call")
-	finalBody["parallel_tool_calls"] = false
-	return finalBody
-}
-
-func forcedFinalResponsesRequest(reqBody map[string]any) map[string]any {
-	finalBody := cloneJSONMap(reqBody)
-	finalBody["input"] = append(forcedFinalResponseInput(finalBody["input"]), map[string]any{
-		"type": "message",
-		"role": "system",
-		"content": []map[string]any{
-			{
-				"type": "input_text",
-				"text": finalAnswerInstructionText,
-			},
-		},
-	})
-	delete(finalBody, "tools")
-	delete(finalBody, "tool_choice")
-	finalBody["parallel_tool_calls"] = false
-	return finalBody
-}
-
-func forcedFinalResponseInput(raw any) []any {
-	input := normalizeResponsesInput(raw)
-	finalInput := make([]any, 0, len(input))
-	for _, rawItem := range input {
-		item, _ := rawItem.(map[string]any)
-		if item == nil {
-			continue
-		}
-
-		switch stringValue(item["type"]) {
-		case "message":
-			finalInput = append(finalInput, rawItem)
-		case "function_call_output":
-			finalInput = append(finalInput, map[string]any{
-				"type": "message",
-				"role": "system",
-				"content": []map[string]any{
-					{
-						"type": "input_text",
-						"text": "Tool result:\n" + fmt.Sprint(item["output"]),
-					},
-				},
-			})
-		}
-	}
-	return finalInput
-}
-
-func normalizeResponsesInput(raw any) []any {
-	switch value := raw.(type) {
-	case nil:
-		return []any{}
-	case []any:
-		return value
-	case []map[string]any:
-		items := make([]any, 0, len(value))
-		for _, item := range value {
-			items = append(items, item)
-		}
-		return items
-	case string:
-		if strings.TrimSpace(value) == "" {
-			return []any{}
-		}
-		return []any{
-			map[string]any{
-				"type": "message",
-				"role": "user",
-				"content": []map[string]any{
-					{
-						"type": "input_text",
-						"text": value,
-					},
-				},
-			},
-		}
-	default:
-		return []any{raw}
-	}
-}
-
-func normalizeResponsesOutputItems(items []any) []any {
-	normalized := make([]any, 0, len(items))
-	for _, rawItem := range items {
-		item, _ := rawItem.(map[string]any)
-		if item == nil {
-			continue
-		}
-
-		if stringValue(item["type"]) != "mcp_call" {
-			normalized = append(normalized, rawItem)
-			continue
-		}
-
-		normalized = append(normalized, map[string]any{
-			"type":      "function_call",
-			"call_id":   firstNonEmptyString(item["call_id"], item["id"]),
-			"name":      item["name"],
-			"arguments": item["arguments"],
-		})
-	}
-	return normalized
-}
-
-func promptMessages(prompt *mcp.GetPromptResult) []any {
-	if prompt == nil {
-		return nil
-	}
-	result := make([]any, 0, len(prompt.Messages))
-	for _, message := range prompt.Messages {
-		textContent, ok := message.Content.(*mcp.TextContent)
-		if !ok || strings.TrimSpace(textContent.Text) == "" {
-			continue
-		}
-		result = append(result, map[string]any{
-			"role":    message.Role,
-			"content": textContent.Text,
-		})
-	}
-	return result
-}
-
-func routedToolDescription(tool *mcp.Tool) string {
-	if tool == nil {
-		return ""
-	}
-	description := strings.TrimSpace(tool.Description)
-	if description == "" {
-		description = fmt.Sprintf("Use the %s tool when it would improve the answer.", tool.Name)
-	}
-	return fmt.Sprintf(
-		"%s Today is %s. Each result includes the source URL and title; cite them inline using standard markdown link syntax, where the label is in square brackets and the URL follows in parentheses, for example [Page title](https://example.com/page). Do not wrap the link in any additional brackets. %s %s %s",
-		description,
-		time.Now().Format(currentDateTimeFormat),
-		citationInstructions,
-		toolOutputWarning,
-		toolEconomyInstructions,
-	)
-}
-
-func chatPromptPrefix(prompt *mcp.GetPromptResult) []any {
-	basePromptMessages := promptMessages(prompt)
-	result := make([]any, 0, len(basePromptMessages)+1)
-	if contextMessage := buildContextMessage(); contextMessage != "" {
-		result = append(result, map[string]any{
-			"role":    "system",
-			"content": contextMessage,
-		})
-	}
-	result = append(result, basePromptMessages...)
-	return result
-}
-
-func promptInputItems(prompt *mcp.GetPromptResult) []any {
-	if prompt == nil {
-		return nil
-	}
-	result := make([]any, 0, len(prompt.Messages))
-	for _, message := range prompt.Messages {
-		textContent, ok := message.Content.(*mcp.TextContent)
-		if !ok || strings.TrimSpace(textContent.Text) == "" {
-			continue
-		}
-		result = append(result, map[string]any{
-			"type": "message",
-			"role": message.Role,
-			"content": []map[string]any{
-				{
-					"type": "input_text",
-					"text": textContent.Text,
-				},
-			},
-		})
-	}
-	return result
-}
-
-func responsePromptPrefix(prompt *mcp.GetPromptResult) []any {
-	basePromptItems := promptInputItems(prompt)
-	result := make([]any, 0, len(basePromptItems)+1)
-	if contextMessage := buildContextMessage(); contextMessage != "" {
-		result = append(result, map[string]any{
-			"type": "message",
-			"role": "system",
-			"content": []map[string]any{
-				{
-					"type": "input_text",
-					"text": contextMessage,
-				},
-			},
-		})
-	}
-	result = append(result, basePromptItems...)
-	return result
-}
-
-func buildContextMessage() string {
-	return fmt.Sprintf(
-		"Current date and time: %s. If the user asks about \"today\", \"latest\", or other time-sensitive topics, interpret them relative to this timestamp and prioritize the freshest tool results.",
-		time.Now().Format(currentDateTimeFormat),
-	)
-}
 
 func isStream(body map[string]any) bool {
 	stream, _ := body["stream"].(bool)
